Replace libxmljs-mt with xml2js
libxmljs(-mt) is a good XML library, but it is extremely annoying to work with simply
because it depends on a native library that needs to be provided, and is a native piece
of code itself.

xml2js is a lot simpler (and particularly doesn't really understand XML namespaces), but
also only comes as a native JS module without ugly dependencies.

For the simple case of reading the subscription ARN from the confirmation URL xml2js is just
fine, and likely AWS won't ever start doing anything with the XML responses -- if the content-type
issue is any indication :)